修改股票插件
