+++ Plugins/Stock/FloatingWnd.cpp
@@ -1043,44 +1043,74 @@ void CFloatingWnd::DrawTimelinePriceCurve(CDC& memDC, const TimelineDrawContext&
 			return ctx.priceChartTop + ctx.priceChartHeight - static_cast<int>(round((price - minPrice) * unitY));
 			};
 
-		const COLORREF navColor = RGB(160, 32, 240);  // 紫色
-		CPen navPen(PS_SOLID, 2, navColor);
-		CPen* pOldPen = memDC.SelectObject(&navPen);
-		bool firstNavPoint = true;
-		int lastNavX = 0, lastNavY = 0;
-		for (int i = 0; i < totalPoints; i++)
+		// 从数据库加载当天的IOPV缓存数据
+		auto navPoints = g_data.GetDbManager().LoadLatestFundNavCache(m_stock_id);
+		if (!navPoints.empty())
 		{
-			const auto& item = timelinePoint[i];
-			if (item.iopv <= 0)
-				continue;
-			int pointX = static_cast<int>(ctx.chartWidth / static_cast<float>(xAxisPts) * i) + static_cast<int>(ctx.chartWidth / static_cast<float>(xAxisPts) / 2);
-			int pointY = priceToY(item.iopv);
-			if (firstNavPoint)
+			// 使用完整分时数据构建时间→索引映射（确保缩放时也能匹配所有IOPV点）
+			// 时间格式：分时数据为"HH:MM:SS"，缓存为"HH:MM"，用前5字符（HH:MM）匹配
+			const auto& fullTimeline = *ctx.fullTimeline;
+			std::map<std::string, int> fullTimeIndexMap;
+			for (int i = 0; i < static_cast<int>(fullTimeline.size()); i++)
 			{
-				memDC.MoveTo(pointX, pointY);
-				lastNavX = pointX;
-				lastNavY = pointY;
-				firstNavPoint = false;
+				// 取前5个字符作为HH:MM
+				std::string hhmm = fullTimeline[i].time.substr(0, 5);
+				fullTimeIndexMap[hhmm] = i;
 			}
-			else
-			{
-				memDC.LineTo(pointX, pointY);
-				lastNavX = pointX;
-				lastNavY = pointY;
-			}
-		}
-		memDC.SelectObject(pOldPen);
 
-		// 在曲线末端标注"净值"文字
-		if (!firstNavPoint)
-		{
-			memDC.SetTextColor(navColor);
-			memDC.SetBkMode(TRANSPARENT);
-			CString navLabel = _T("净值");
-			CSize sz = memDC.GetTextExtent(navLabel);
-			int labelX = lastNavX + g_data.RDPI(3);
-			int labelY = lastNavY - sz.cy / 2;
-			memDC.TextOut(labelX, labelY, navLabel);
+			// 将IOPV缓存数据映射到完整分时数据的索引
+			std::map<int, double> iopvByIndex;  // fullTimeline索引 → iopv值
+			for (const auto& nav : navPoints)
+			{
+				// 缓存时间已经是HH:MM格式
+				auto it = fullTimeIndexMap.find(nav.time);
+				if (it != fullTimeIndexMap.end())
+					iopvByIndex[it->second] = nav.iopv;
+			}
+
+			// 加上当前实时IOPV（如果有）
+			if (ctx.realtimeData.iopv > 0 && !fullTimeline.empty())
+			{
+				int lastIdx = static_cast<int>(fullTimeline.size()) - 1;
+				iopvByIndex[lastIdx] = ctx.realtimeData.iopv;
+			}
+
+			if (!iopvByIndex.empty())
+			{
+				const COLORREF navColor = RGB(160, 32, 240);  // 紫色
+				CPen navPen(PS_SOLID, 1, navColor);
+				CPen* pOldPen = memDC.SelectObject(&navPen);
+				bool firstNavPoint = true;
+
+				// 计算startIndex对应的完整数据偏移
+				int startIdx = ctx.startIndex;
+				int visCount = ctx.visibleCount;
+
+				for (const auto& kv : iopvByIndex)
+				{
+					int fullIdx = kv.first;
+					double iopvVal = kv.second;
+
+					// 只绘制可见范围内的点
+					if (fullIdx < startIdx || fullIdx >= startIdx + visCount)
+						continue;
+
+					// 计算在可见区域内的相对索引
+					int relIdx = fullIdx - startIdx;
+					int pointX = static_cast<int>(ctx.chartWidth / static_cast<float>(xAxisPts) * relIdx) + static_cast<int>(ctx.chartWidth / static_cast<float>(xAxisPts) / 2);
+					int pointY = priceToY(iopvVal);
+					if (firstNavPoint)
+					{
+						memDC.MoveTo(pointX, pointY);
+						firstNavPoint = false;
+					}
+					else
+					{
+						memDC.LineTo(pointX, pointY);
+					}
+				}
+				memDC.SelectObject(pOldPen);
+			}
 		}
 	}
 
@@ -2740,7 +2770,7 @@ void CFloatingWnd::OnPaint()
 			const double DIV_COUNT = 6.0;
 			const double MIN_STEP = 0.001;
 			double axisMin, axisMax, niceStep;
-			CalcNiceAxisRange(visMin, visMax, DIV_COUNT, MIN_STEP, axisMin, axisMax, niceStep);
+			CStockIndicator::CalcNiceAxisRange(visMin, visMax, DIV_COUNT, MIN_STEP, axisMin, axisMax, niceStep);
 			ctx.maxPrice = axisMax;
 			ctx.minPrice = axisMin;
 			ctx.niceStep = niceStep;
@@ -2802,7 +2832,7 @@ void CFloatingWnd::OnPaint()
 		else if (!timelinePoint.empty())
 		{
 			// 先基于完整分时数据计算MA，避免缩放/拖动后只用可见区间导致均线与其他APP不一致
-			CalcAllRollingAvgPrices(timelinePoint);
+			CStockIndicator::CalcAllRollingAvgPrices(timelinePoint);
 
 			// 计算可见范围：m_timelineVisibleCount控制缩放，m_timelineScrollOffset控制拖动
 			int totalPoints = static_cast<int>(timelinePoint.size());
@@ -2946,7 +2976,7 @@ void CFloatingWnd::OnPaint()
 				const double DIV_COUNT = 6.0;
 				const double MIN_STEP = 0.001;
 				double axisMin, axisMax, niceStep;
-				CalcNiceAxisRange(visMin, visMax, DIV_COUNT, MIN_STEP, axisMin, axisMax, niceStep);
+				CStockIndicator::CalcNiceAxisRange(visMin, visMax, DIV_COUNT, MIN_STEP, axisMin, axisMax, niceStep);
 
 				ctx.maxPrice = axisMax;
 				ctx.minPrice = axisMin;
@@ -3264,18 +3294,22 @@ void CFloatingWnd::OnPaint()
 					RegResult trend = (!m_isKLineMode)
 						? g_data.Get1MinAvgTrend(std::wstring(m_stock_id))
 						: g_data.Get5MinAvgTrend(std::wstring(m_stock_id));
-					if (trend.valid && (trend.r2 >= 0.2 || std::abs(trend.slope) >= 0.01))
+					trendArrowStr = _T("|"); // 默认竖线
+					if (trend.valid)
 					{
-						// 箭头强度：低1个、中2个、高3个
-						int arrowCount = 1;
-						if (trend.r2 >= 0.7)
-							arrowCount = 3;
-						else if (trend.r2 >= 0.55)
-							arrowCount = 2;
-						if (trend.slope > 0)
-							trendArrowStr = CString(_T('↑'), arrowCount);
-						else
-							trendArrowStr = CString(_T('↓'), arrowCount);
+						if (trend.r2 >= 0.2 || std::abs(trend.slope) >= 0.01)
+						{
+							// 箭头强度：低1个、中2个、高3个
+							int arrowCount = 1;
+							if (trend.r2 >= 0.7)
+								arrowCount = 3;
+							else if (trend.r2 >= 0.55)
+								arrowCount = 2;
+							if (trend.slope > 0)
+								trendArrowStr = CString(_T('↑'), arrowCount);
+							else
+								trendArrowStr = CString(_T('↓'), arrowCount);
+						}
 					}
 				}
 			}
@@ -3499,8 +3533,22 @@ void CFloatingWnd::OnPaint()
 					// 切换到固定字体绘制均幅区域
 					CFont* pPrevFont = memDC.SelectObject(&avgFont);
 
-					// 三等分120像素区域
+					// 120像素区域三等分
 					int thirdWidth = avgAreaWidth / 3;
+
+					// 趋势指示器（120像素左侧，间隔2像素，无背景）
+					if (!trendArrowStr.IsEmpty())
+					{
+						TCHAR firstChar = trendArrowStr.GetAt(0);
+						if (firstChar == _T('↑'))
+							memDC.SetTextColor(COLOR_RED_UP);
+						else if (firstChar == _T('↓'))
+							memDC.SetTextColor(COLOR_GREEN_DOWN);
+						else
+							memDC.SetTextColor(RGB(0, 0, 0)); // 竖线用黑色
+						int trendX = avgAreaX - GAP - memDC.GetTextExtent(trendArrowStr).cx;
+						memDC.TextOut(trendX, avgAreaY + g_data.RDPI(2), trendArrowStr);
+					}
 
 					// 最小值（白色文字）
 					memDC.SetTextColor(RGB(255, 255, 255));
@@ -3514,15 +3562,6 @@ void CFloatingWnd::OnPaint()
 					// 最大值（白色文字）
 					int maxX = avgAreaX + thirdWidth * 2 + (thirdWidth - memDC.GetTextExtent(maxAvgValueStr).cx) / 2;
 					memDC.TextOut(maxX, avgAreaY + g_data.RDPI(2), maxAvgValueStr);
-
-					// 绘制趋势箭头（紧接120像素区域右侧，仍在固定字体下）
-					if (!trendArrowStr.IsEmpty())
-					{
-						bool isUp = trendArrowStr.GetAt(0) == _T('↑');
-						memDC.SetTextColor(isUp ? COLOR_RED_UP : COLOR_GREEN_DOWN);
-						int arrowX = avgAreaX + avgAreaWidth + GAP;
-						memDC.TextOut(arrowX, avgAreaY + g_data.RDPI(2), trendArrowStr);
-					}
 
 					// 恢复之前的字体
 					memDC.SelectObject(pPrevFont);
@@ -3696,651 +3735,12 @@ void CFloatingWnd::OnPaint()
 	memDC.SelectObject(pOldBitmap);
 }
 
-// ========== MACD指标计算与绘制 ==========
-
-// 为每个分时数据点计算MA5/MA10/MA20滚动均价（滑动窗口）
-void CFloatingWnd::CalcAllRollingAvgPrices(std::vector<STOCK::TimelinePoint>& timelinePoint)
-{
-	int n = static_cast<int>(timelinePoint.size());
-	if (n == 0)
-		return;
-
-	// 计算每个窗口的滚动均价，使用滑动窗口避免重复求和
-	auto calcWindow = [&](int windowSize, int fieldOffset) {
-		STOCK::Amount sumAmount = 0;
-		STOCK::Volume sumVolume = 0;
-		for (int i = 0; i < n; i++)
-		{
-			sumAmount += timelinePoint[i].amount;
-			sumVolume += timelinePoint[i].volume;
-			if (i >= windowSize)
-			{
-				sumAmount -= timelinePoint[i - windowSize].amount;
-				sumVolume -= timelinePoint[i - windowSize].volume;
-			}
-			if (sumVolume > 0)
-			{
-				STOCK::Price maVal = sumAmount / sumVolume;
-				switch (fieldOffset)
-				{
-				case 5: timelinePoint[i].ma5 = maVal; break;
-				case 10: timelinePoint[i].ma10 = maVal; break;
-				case 20: timelinePoint[i].ma20 = maVal; break;
-				}
-			}
-		}
-		};
-
-	calcWindow(5, 5);
-	calcWindow(10, 10);
-	calcWindow(20, 20);
-}
-
-// 计算滚动均价：最近nMinutes分钟的成交额/成交量（单次查询）
-STOCK::Price CFloatingWnd::CalcRollingAvgPrice(const std::vector<STOCK::TimelinePoint>& timelinePoint, int nMinutes)
-{
-	int n = static_cast<int>(timelinePoint.size());
-	if (n == 0 || nMinutes <= 0)
-		return 0;
-
-	int startIdx = max(0, n - nMinutes);
-	STOCK::Amount sumAmount = 0;
-	STOCK::Volume sumVolume = 0;
-	for (int i = startIdx; i < n; i++)
-	{
-		sumAmount += timelinePoint[i].amount;
-		sumVolume += timelinePoint[i].volume;
-	}
-	if (sumVolume == 0)
-		return 0;
-	return sumAmount / sumVolume;
-}
-
-double CFloatingWnd::CalcNiceStep(double range, double divCount, double minStep)
-{
-	if (range <= 0) range = minStep;
-	double rawStep = range / divCount;
-	if (rawStep <= 0) rawStep = minStep;
-	double mag = pow(10.0, floor(log10(rawStep)));
-	double norm = rawStep / mag;
-	// norm ∈ [1,10)，映射到1~10的整数步长，平滑过渡
-	static const double thresholds[] = { 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0 };
-	double niceNorm = 10.0;
-	for (int i = 0; i < _countof(thresholds); i++)
-	{
-		if (norm <= thresholds[i] + 1e-9)
-		{
-			niceNorm = thresholds[i];
-			break;
-		}
-	}
-	return (std::max)(niceNorm * mag, minStep);
-}
-
-void CFloatingWnd::CalcNiceAxisRange(double visMin, double visMax, double divCount, double minStep, double& outAxisMin, double& outAxisMax, double& outNiceStep)
-{
-	double range = visMax - visMin;
-	double niceStep = CalcNiceStep(range, divCount, minStep);
-
-	// 先对齐axisMin到niceStep的整数倍，确保数据完全包含在轴范围内
-	double axisMin = floor((visMin + niceStep * 1e-9) / niceStep) * niceStep;
-	double axisMax = axisMin + divCount * niceStep;
-
-	// 如果axisRange不够包含所有数据，增加等分数
-	while (axisMax < visMax - 1e-9)
-	{
-		axisMax += niceStep;
-	}
-
-	// 三位小数精度截断（与显示格式一致）
-	axisMin = round(axisMin * 1000.0) / 1000.0;
-	axisMax = round(axisMax * 1000.0) / 1000.0;
-
-	// 股价非负约束
-	if (axisMin < 0)
-	{
-		axisMin = 0;
-		axisMax = round((axisMin + ceil((visMax + niceStep * 1e-9) / niceStep) * niceStep) * 1000.0) / 1000.0;
-	}
-
-	outAxisMin = axisMin;
-	outAxisMax = axisMax;
-	outNiceStep = niceStep;
-}
-
-void CFloatingWnd::CalcNiceAxisRangeSymmetric(double visMin, double visMax, double divCount, double minStep, double& outMin, double& outMax, double& outNiceStep)
-{
-	double range = visMax - visMin;
-	double niceStep = CalcNiceStep(range, divCount, minStep);
-
-	double centerPrice = (visMin + visMax) / 2.0;
-	double halfAxisRange = (divCount / 2.0) * niceStep;
-	outMin = centerPrice - halfAxisRange;
-	outMax = centerPrice + halfAxisRange;
-	outNiceStep = niceStep;
-}
-
-std::vector<CFloatingWnd::MACDData> CFloatingWnd::CalculateTimelineMACD(const std::vector<STOCK::TimelinePoint>& timelinePoint)
-{
-	std::vector<MACDData> result;
-	int n = static_cast<int>(timelinePoint.size());
-	if (n == 0)
-		return result;
-
-	result.resize(n);
-
-	// 提取收盘价（分时价格）
-	std::vector<double> closes(n);
-	for (int i = 0; i < n; i++)
-		closes[i] = timelinePoint[i].price;
-
-	// 通达信初始化方式：第一根收盘价作为EMA初始值，提前生成指标
-	// 随着数据量增加，数值逐步收敛到标准值
-
-	// 计算EMA12：第一根用收盘价初始化，后续迭代
-	std::vector<double> ema12(n);
-	ema12[0] = closes[0];
-	for (int i = 1; i < n; i++)
-		ema12[i] = closes[i] * 2.0 / 13.0 + ema12[i - 1] * 11.0 / 13.0;
-
-	// 计算EMA26：第一根用收盘价初始化，后续迭代
-	std::vector<double> ema26(n);
-	ema26[0] = closes[0];
-	for (int i = 1; i < n; i++)
-		ema26[i] = closes[i] * 2.0 / 27.0 + ema26[i - 1] * 25.0 / 27.0;
-
-	// 计算DIF = EMA12 - EMA26
-	std::vector<double> dif(n);
-	for (int i = 0; i < n; i++)
-		dif[i] = ema12[i] - ema26[i];
-
-	// 计算DEA（DIF的9周期EMA）：第一根DIF作为初始值，后续迭代
-	std::vector<double> dea(n);
-	dea[0] = dif[0];
-	for (int i = 1; i < n; i++)
-		dea[i] = dif[i] * 2.0 / 11.0 + dea[i - 1] * 9.0 / 11.0;
-
-	// 计算MACD柱 = (DIF - DEA) * 2，所有数据点均有效
-	for (int i = 0; i < n; i++)
-	{
-		result[i].dif = dif[i];
-		result[i].dea = dea[i];
-		result[i].bar = (dif[i] - dea[i]) * 2.0;
-		result[i].valid = true;
-	}
-
-	return result;
-}
-
-std::vector<CFloatingWnd::MACDData> CFloatingWnd::CalculateKLineMACD(const std::vector<STOCK::KLinePoint>& klineData)
-{
-	std::vector<MACDData> result;
-	int n = static_cast<int>(klineData.size());
-	if (n == 0)
-		return result;
-
-	result.resize(n);
-
-	// 提取收盘价
-	std::vector<double> closes(n);
-	for (int i = 0; i < n; i++)
-		closes[i] = klineData[i].close;
-
-	// 计算EMA12
-	std::vector<double> ema12(n);
-	ema12[0] = closes[0];
-	for (int i = 1; i < n; i++)
-		ema12[i] = closes[i] * 2.0 / 13.0 + ema12[i - 1] * 11.0 / 13.0;
-
-	// 计算EMA26
-	std::vector<double> ema26(n);
-	ema26[0] = closes[0];
-	for (int i = 1; i < n; i++)
-		ema26[i] = closes[i] * 2.0 / 27.0 + ema26[i - 1] * 25.0 / 27.0;
-
-	// 计算DIF = EMA12 - EMA26
-	std::vector<double> dif(n);
-	for (int i = 0; i < n; i++)
-		dif[i] = ema12[i] - ema26[i];
-
-	// 计算DEA（DIF的9周期EMA）
-	std::vector<double> dea(n);
-	dea[0] = dif[0];
-	for (int i = 1; i < n; i++)
-		dea[i] = dif[i] * 2.0 / 11.0 + dea[i - 1] * 9.0 / 11.0;
-
-	// 计算MACD柱 = (DIF - DEA) * 2
-	for (int i = 0; i < n; i++)
-	{
-		result[i].dif = dif[i];
-		result[i].dea = dea[i];
-		result[i].bar = (dif[i] - dea[i]) * 2.0;
-		result[i].valid = true;
-	}
-
-	return result;
-}
-
-std::vector<CFloatingWnd::MACDCrossSignal> CFloatingWnd::DetectMACDCross(const std::vector<MACDData>& macdData)
-{
-	std::vector<MACDCrossSignal> signals(macdData.size(), MACDCrossSignal::None);
-	if (macdData.size() < 2)
-		return signals;
-
-	const double epsilon = 1e-6; // 粘合阈值：差值小于此值视为未交叉
-	const int MIN_CROSS_GAP = 10; // 两次交叉之间最少间隔10根K线，防止0轴附近频繁缠绕
-	int lastCrossIdx = -MIN_CROSS_GAP; // 上次交叉位置
-
-	for (size_t i = 1; i < macdData.size(); i++)
-	{
-		if (!macdData[i].valid || !macdData[i - 1].valid)
-			continue;
-
-		double difPrev = macdData[i - 1].dif;
-		double deaPrev = macdData[i - 1].dea;
-		double difNow = macdData[i].dif;
-		double deaNow = macdData[i].dea;
-
-		// 粘合状态：当前DIF与DEA差值过小，不触发新叉
-		if (std::abs(difNow - deaNow) < epsilon)
-			continue;
-
-		// 金叉：上一时刻DIF < DEA，当前时刻DIF > DEA
-		if (difPrev < deaPrev && difNow > deaNow)
-		{
-			if (static_cast<int>(i) - lastCrossIdx >= MIN_CROSS_GAP)
-			{
-				signals[i] = MACDCrossSignal::GoldenCross;
-				lastCrossIdx = static_cast<int>(i);
-			}
-		}
-		// 死叉：上一时刻DIF > DEA，当前时刻DIF < DEA
-		else if (difPrev > deaPrev && difNow < deaNow)
-		{
-			if (static_cast<int>(i) - lastCrossIdx >= MIN_CROSS_GAP)
-			{
-				signals[i] = MACDCrossSignal::DeathCross;
-				lastCrossIdx = static_cast<int>(i);
-			}
-		}
-	}
-
-	return signals;
-}
-
-CFloatingWnd::MACDCrossSignal CFloatingWnd::GetLatestMACDCross(const std::vector<MACDData>& macdData)
-{
-	auto signals = DetectMACDCross(macdData);
-	for (int i = static_cast<int>(signals.size()) - 1; i >= 0; i--)
-	{
-		if (signals[i] != MACDCrossSignal::None)
-			return signals[i];
-	}
-	return MACDCrossSignal::None;
-}
-
-// ========== T+0日内买卖点实时判定 ==========
-
-CSignalAnalyzer::T0Signal CFloatingWnd::DetectBuySignal(const std::vector<STOCK::TimelinePoint>& timelinePoint, const std::vector<MACDData>& macdData)
-{
-	CSignalAnalyzer::T0Signal signal;
-	signal.valid = false;
-	signal.isBuy = true;
-	signal.strength = 0;
-	signal.price = 0;
-
-	int n = static_cast<int>(timelinePoint.size());
-	int m = static_cast<int>(macdData.size());
-	if (n < 10 || m < 10)
-		return signal;
-
-	int last = n - 1;
-	signal.price = timelinePoint[last].price;
-	signal.time = CString(timelinePoint[last].time.c_str());
-
-	// ========== 量能基准：前三根成交量均值 ==========
-	long long volAvg3 = 0;
-	{
-		int volCount = min(3, last);
-		for (int i = last - volCount; i < last; i++)
-			volAvg3 += timelinePoint[i].volume;
-		if (volCount > 0) volAvg3 /= volCount;
-	}
-	bool isShrinking = (volAvg3 > 0 && timelinePoint[last].volume < volAvg3);           // 缩量：当前量 < 前三根均值
-	bool isExpanding = (volAvg3 > 0 && timelinePoint[last].volume > volAvg3 * 3 / 2);    // 放量：当前量 > 前三根均值的1.5倍
-
-	int strength = 0;
-	CString reasons;
-
-	// ========== 避坑规则：放量大跌不抄底 ==========
-	{
-		int lookback = min(5, n - 1);
-		bool heavyDecline = true;
-		for (int i = last - lookback + 1; i <= last; i++)
-		{
-			if (timelinePoint[i].price >= timelinePoint[i - 1].price)
-			{
-				heavyDecline = false;
-				break;
-			}
-		}
-		// 持续下跌且量能递增 = 放量杀跌
-		if (heavyDecline)
-		{
-			bool volIncreasing = true;
-			for (int i = last - lookback + 2; i <= last; i++)
-			{
-				if (timelinePoint[i].volume < timelinePoint[i - 1].volume)
-				{
-					volIncreasing = false;
-					break;
-				}
-			}
-			if (volIncreasing)
-				return signal; // 放量杀跌，坚决不抄底
-		}
-	}
-
-	// ========== 避坑规则：无量阴跌不急于接盘 ==========
-	{
-		int lookback = min(10, n - 1);
-		bool steadyDecline = true;
-		for (int i = last - lookback + 1; i <= last; i++)
-		{
-			if (timelinePoint[i].price >= timelinePoint[i - 1].price)
-			{
-				steadyDecline = false;
-				break;
-			}
-		}
-		bool lowVolume = isShrinking;
-		bool macdStillDown = macdData[last].valid && macdData[last].dif < 0 && macdData[last].bar < 0;
-		if (steadyDecline && lowVolume && macdStillDown)
-			return signal; // 无量阴跌，MACD仍下行
-	}
-
-	// ========== 1. 量柱信号：缩量下跌 + 首根红量柱（需量能确认） ==========
-	{
-		int lookback = min(10, n - 1);
-		int greenCount = 0;
-		bool shrinking = true;
-		long long prevGreenVol = 0;
-
-		for (int i = last - lookback + 1; i <= last; i++)
-		{
-			bool isGreen = (timelinePoint[i].price < timelinePoint[i - 1].price);
-			if (isGreen)
-			{
-				greenCount++;
-				if (prevGreenVol > 0 && timelinePoint[i].volume > prevGreenVol)
-					shrinking = false;
-				prevGreenVol = timelinePoint[i].volume;
-			}
-		}
-
-		// 最新一根是否为红柱（上涨），且之前有连续下跌
-		bool firstRed = (timelinePoint[last].price > timelinePoint[last - 1].price) && greenCount >= 3;
-
-		if (shrinking && greenCount >= 3 && isShrinking)
-		{
-			strength++;
-			reasons += _T("缩量下跌抛压耗尽 ");
-		}
-		// 首根红量柱需要放量确认，否则可能是假反弹
-		if (firstRed && isExpanding)
-		{
-			strength++;
-			reasons += _T("放量红柱资金承接 ");
-		}
-	}
-
-	// ========== 2. MACD信号 ==========
-	if (macdData[last].valid)
-	{
-		// 绿柱缩短
-		if (macdData[last].bar < 0)
-		{
-			int shrkCount = 0;
-			for (int i = last; i > max(0, last - 5); i--)
-			{
-				if (macdData[i].valid && macdData[i].bar < 0 &&
-					std::abs(macdData[i].bar) <= std::abs(macdData[i - 1].bar))
-					shrkCount++;
-			}
-			if (shrkCount >= 3)
-			{
-				strength++;
-				reasons += _T("MACD绿柱缩短 ");
-			}
-		}
-
-		// 底背离检测：价格创新低，MACD未创新低
-		int lookback = min(30, n - 1);
-		int low1Idx = -1, low2Idx = -1;
-		for (int i = last - lookback + 1; i < last; i++)
-		{
-			if (i > 0 && i < n - 1 &&
-				timelinePoint[i].price < timelinePoint[i - 1].price &&
-				timelinePoint[i].price < timelinePoint[i + 1].price)
-			{
-				if (low1Idx < 0)
-					low1Idx = i;
-				else if (i > low1Idx + 3 && low2Idx < 0)
-					low2Idx = i;
-			}
-		}
-		if (low1Idx >= 0 && low2Idx >= 0 &&
-			timelinePoint[low2Idx].price < timelinePoint[low1Idx].price &&
-			macdData[low2Idx].valid && macdData[low1Idx].valid &&
-			macdData[low2Idx].dif > macdData[low1Idx].dif)
-		{
-			strength += 2;
-			reasons += _T("MACD底背离 ");
-		}
-	}
-
-	// ========== 3. 股价支撑（需量能确认） ==========
-	{
-		// 回踩均价支撑：需要缩量回踩才有效
-		double avgPrice = timelinePoint[last].averagePrice;
-		double price = timelinePoint[last].price;
-		if (avgPrice > 0 && std::abs(price - avgPrice) / avgPrice < 0.003 && isShrinking)
-		{
-			strength++;
-			reasons += _T("缩量回踩均价支撑 ");
-		}
-		// 分时底部企稳：需要放量止跌才可靠
-		if (last >= 2 &&
-			timelinePoint[last].price >= timelinePoint[last - 1].price &&
-			timelinePoint[last - 1].price < timelinePoint[last - 2].price &&
-			isExpanding)
-		{
-			strength++;
-			reasons += _T("放量底部企稳 ");
-		}
-	}
-
-	signal.strength = min(3, strength);
-	signal.reason = reasons;
-	signal.valid = (strength >= 2);
-	return signal;
-}
-
-CSignalAnalyzer::T0Signal CFloatingWnd::DetectSellSignal(const std::vector<STOCK::TimelinePoint>& timelinePoint, const std::vector<MACDData>& macdData)
-{
-	CSignalAnalyzer::T0Signal signal;
-	signal.valid = false;
-	signal.isBuy = false;
-	signal.strength = 0;
-	signal.price = 0;
-
-	int n = static_cast<int>(timelinePoint.size());
-	int m = static_cast<int>(macdData.size());
-	if (n < 10 || m < 10)
-		return signal;
-
-	int last = n - 1;
-	signal.price = timelinePoint[last].price;
-	signal.time = CString(timelinePoint[last].time.c_str());
-
-	// ========== 量能基准：前三根成交量均值 ==========
-	long long volAvg3 = 0;
-	{
-		int volCount = min(3, last);
-		for (int i = last - volCount; i < last; i++)
-			volAvg3 += timelinePoint[i].volume;
-		if (volCount > 0) volAvg3 /= volCount;
-	}
-	bool isShrinking = (volAvg3 > 0 && timelinePoint[last].volume < volAvg3);           // 缩量：当前量 < 前三根均值
-	bool isExpanding = (volAvg3 > 0 && timelinePoint[last].volume > volAvg3 * 3 / 2);    // 放量：当前量 > 前三根均值的1.5倍
-
-	int strength = 0;
-	CString reasons;
-
-	// ========== 避坑规则：放量大涨不卖出 ==========
-	{
-		int lookback = min(5, n - 1);
-		bool strongRise = true;
-		for (int i = last - lookback + 1; i <= last; i++)
-		{
-			if (timelinePoint[i].price <= timelinePoint[i - 1].price)
-			{
-				strongRise = false;
-				break;
-			}
-		}
-		if (strongRise)
-		{
-			bool volIncreasing = true;
-			for (int i = last - lookback + 2; i <= last; i++)
-			{
-				if (timelinePoint[i].volume < timelinePoint[i - 1].volume)
-				{
-					volIncreasing = false;
-					break;
-				}
-			}
-			if (volIncreasing)
-				return signal; // 放量大涨，不卖出
-		}
-	}
-
-	// ========== 1. 量柱信号：放量冲高 + 量能萎缩（需量能确认） ==========
-	{
-		// 检查近期是否有放量冲高（使用前三根均值判定放量）
-		int lookback = min(10, n - 1);
-		bool hadSpike = false;
-		int spikeIdx = -1;
-		for (int i = last - lookback + 1; i <= last; i++)
-		{
-			// 计算该点的前三根均值
-			long long localAvg3 = 0;
-			int vc = min(3, i);
-			for (int j = i - vc; j < i; j++)
-				localAvg3 += timelinePoint[j].volume;
-			if (vc > 0) localAvg3 /= vc;
-			if (localAvg3 > 0 && timelinePoint[i].volume > localAvg3 * 3 / 2)
-			{
-				hadSpike = true;
-				spikeIdx = i;
-				break;
-			}
-		}
-		// 放量后量能萎缩，且当前价格未明显回落（仍在高位）
-		bool priceStillHigh = (hadSpike && spikeIdx >= 0 &&
-			timelinePoint[last].price >= timelinePoint[spikeIdx].price * 0.995);
-		bool shrinkingNow = false;
-		if (hadSpike && last > spikeIdx + 1 &&
-			timelinePoint[last].volume < timelinePoint[last - 1].volume &&
-			isShrinking && priceStillHigh)
-			shrinkingNow = true;
-
-		if (hadSpike && shrinkingNow)
-		{
-			strength++;
-			reasons += _T("放量冲高后缩量回落 ");
-		}
-	}
-
-	// ========== 2. MACD信号 ==========
-	if (macdData[last].valid)
-	{
-		// 红柱缩短
-		if (macdData[last].bar > 0)
-		{
-			int shrkCount = 0;
-			for (int i = last; i > max(0, last - 5); i--)
-			{
-				if (macdData[i].valid && macdData[i].bar > 0 &&
-					macdData[i].bar <= macdData[i - 1].bar)
-					shrkCount++;
-			}
-			if (shrkCount >= 3)
-			{
-				strength++;
-				reasons += _T("MACD红柱缩短 ");
-			}
-		}
-
-		// 死叉检测
-		auto crossSignals = DetectMACDCross(macdData);
-		if (last < static_cast<int>(crossSignals.size()) && crossSignals[last] == MACDCrossSignal::DeathCross)
-		{
-			strength++;
-			reasons += _T("MACD死叉 ");
-		}
-
-		// 顶背离检测：价格创新高，MACD未创新高
-		int lookback = min(30, n - 1);
-		int high1Idx = -1, high2Idx = -1;
-		for (int i = last - lookback + 1; i < last; i++)
-		{
-			if (i > 0 && i < n - 1 &&
-				timelinePoint[i].price > timelinePoint[i - 1].price &&
-				timelinePoint[i].price > timelinePoint[i + 1].price)
-			{
-				if (high1Idx < 0)
-					high1Idx = i;
-				else if (i > high1Idx + 3 && high2Idx < 0)
-					high2Idx = i;
-			}
-		}
-		if (high1Idx >= 0 && high2Idx >= 0 &&
-			timelinePoint[high2Idx].price > timelinePoint[high1Idx].price &&
-			macdData[high2Idx].valid && macdData[high1Idx].valid &&
-			macdData[high2Idx].dif < macdData[high1Idx].dif)
-		{
-			strength += 2;
-			reasons += _T("MACD顶背离 ");
-		}
-	}
-
-	// ========== 3. 股价压力（需量能确认） ==========
-	{
-		// 遇均价压力：需要放量冲高遇阻才有效
-		double avgPrice = timelinePoint[last].averagePrice;
-		double price = timelinePoint[last].price;
-		if (avgPrice > 0 && price > avgPrice && std::abs(price - avgPrice) / avgPrice < 0.005 && isExpanding)
-		{
-			strength++;
-			reasons += _T("放量遇均价压力 ");
-		}
-		// 冲高滞涨：放量冲高后回落
-		if (last >= 2 &&
-			timelinePoint[last].price <= timelinePoint[last - 1].price &&
-			timelinePoint[last - 1].price > timelinePoint[last - 2].price &&
-			isShrinking)
-		{
-			strength++;
-			reasons += _T("冲高缩量滞涨 ");
-		}
-	}
-
-	signal.strength = min(3, strength);
-	signal.reason = reasons;
-	signal.valid = (strength >= 2);
-	return signal;
-}
+// ========== MACD指标绘制 ==========
+// 注：以下指标计算函数已移至 CStockIndicator 类（StockIndicator.h/cpp）：
+//   CalcAllRollingAvgPrices/CalcRollingAvgPrice/CalcNiceStep/CalcNiceAxisRange/
+//   CalcNiceAxisRangeSymmetric/CalculateTimelineMACD/CalculateKLineMACD/
+//   DetectMACDCross/GetLatestMACDCross/DetectBuySignal/DetectSellSignal
+// CFloatingWnd 仅保留绘制逻辑。
 
 void CFloatingWnd::DrawMACDChart(CDC& memDC, int x, int y, int width, int height, const std::vector<STOCK::TimelinePoint>& timelinePoint, const std::vector<MACDData>& macdData, int startIndex /* = 0 */, int visibleCount /* = -1 */, int xAxisPoints /* = 0 */)
 {
@@ -4458,7 +3858,7 @@ void CFloatingWnd::DrawMACDChart(CDC& memDC, int x, int y, int width, int height
 	memDC.SelectObject(pOldPen);
 
 	// 绘制金叉死叉标记
-	auto crossSignals = DetectMACDCross(macdData);
+	auto crossSignals = CStockIndicator::DetectMACDCross(macdData);
 	const int dotRadius = g_data.RDPI(3);
 	const int labelOffset = 0;
 	int oldBkMode = memDC.SetBkMode(TRANSPARENT);
@@ -4615,7 +4015,7 @@ void CFloatingWnd::DrawMACDChart(CDC& memDC, int x, int y, int width, int height
 void CFloatingWnd::DrawTimelineMACDSection(CDC& memDC, const TimelineDrawContext& ctx)
 {
 	const auto& timelinePoint = *ctx.timelinePoint;
-	auto macdData = CalculateTimelineMACD(timelinePoint);
+	auto macdData = CStockIndicator::CalculateTimelineMACD(timelinePoint);
 
 	DrawMACDChart(memDC, 0, ctx.macdChartTop, ctx.chartWidth, ctx.macdChartHeight, timelinePoint, macdData, 0, -1, ctx.xAxisPoints);
 
@@ -4666,7 +4066,7 @@ void CFloatingWnd::DrawTimelineKDJSection(CDC& memDC, const TimelineDrawContext&
 	const auto& timelinePoint = *ctx.timelinePoint;
 	// 使用完整分时数据计算KDJ，避免缩放/拖动时前N-1个点无有效RSV
 	const auto& fullData = ctx.fullTimeline ? *ctx.fullTimeline : timelinePoint;
-	auto kdjData = CalculateTimelineKDJ(fullData);
+	auto kdjData = CStockIndicator::CalculateTimelineKDJ(fullData);
 
 	DrawTimelineKDJChart(memDC, 0, ctx.macdChartTop, ctx.chartWidth, ctx.macdChartHeight, timelinePoint, kdjData, ctx.startIndex, ctx.xAxisPoints);
 
@@ -4802,163 +4202,8 @@ void CFloatingWnd::DrawTimelineKDJChart(CDC& memDC, int x, int y, int width, int
 	memDC.SelectObject(pOldPen);
 }
 
-// ========== W&R威廉指标 ==========
-
-std::vector<CFloatingWnd::WRData> CFloatingWnd::CalculateTimelineWR(const std::vector<STOCK::TimelinePoint>& timelinePoint, int period1 /* = 6 */, int period2 /* = 14 */)
-{
-	std::vector<WRData> result;
-	result.reserve(timelinePoint.size());
-	if (timelinePoint.empty() || period1 <= 0 || period2 <= 0)
-		return result;
-
-	int maxPeriod = (std::max)(period1, period2);
-
-	for (int i = 0; i < static_cast<int>(timelinePoint.size()); i++)
-	{
-		WRData wr;
-		wr.wr1 = 0;
-		wr.wr2 = 0;
-		wr.valid = false;
-
-		// 计算WR1（短期，如N=6）
-		if (i >= period1 - 1)
-		{
-			STOCK::Price highest = 0;
-			STOCK::Price lowest = (std::numeric_limits<STOCK::Price>::max)();
-			bool hasValid = false;
-			for (int j = i - period1 + 1; j <= i; j++)
-			{
-				STOCK::Price price = timelinePoint[j].price;
-				if (price > 0)
-				{
-					highest = (std::max)(highest, price);
-					lowest = (std::min)(lowest, price);
-					hasValid = true;
-				}
-			}
-			if (hasValid && highest > lowest && timelinePoint[i].price > 0)
-			{
-				wr.wr1 = (static_cast<double>(highest) - timelinePoint[i].price) / (highest - lowest) * 100.0;
-				if (wr.wr1 < 0) wr.wr1 = 0;
-				if (wr.wr1 > 100) wr.wr1 = 100;
-			}
-			else if (hasValid && highest == lowest && timelinePoint[i].price > 0)
-			{
-				wr.wr1 = 0;  // 最高=最低时WR=0
-			}
-		}
-
-		// 计算WR2（长期，如N=14）
-		if (i >= period2 - 1)
-		{
-			STOCK::Price highest = 0;
-			STOCK::Price lowest = (std::numeric_limits<STOCK::Price>::max)();
-			bool hasValid = false;
-			for (int j = i - period2 + 1; j <= i; j++)
-			{
-				STOCK::Price price = timelinePoint[j].price;
-				if (price > 0)
-				{
-					highest = (std::max)(highest, price);
-					lowest = (std::min)(lowest, price);
-					hasValid = true;
-				}
-			}
-			if (hasValid && highest > lowest && timelinePoint[i].price > 0)
-			{
-				wr.wr2 = (static_cast<double>(highest) - timelinePoint[i].price) / (highest - lowest) * 100.0;
-				if (wr.wr2 < 0) wr.wr2 = 0;
-				if (wr.wr2 > 100) wr.wr2 = 100;
-			}
-			else if (hasValid && highest == lowest && timelinePoint[i].price > 0)
-			{
-				wr.wr2 = 0;
-			}
-		}
-
-		wr.valid = (i >= maxPeriod - 1);
-		result.push_back(wr);
-	}
-
-	return result;
-}
-
-std::vector<CFloatingWnd::WRData> CFloatingWnd::CalculateKLineWR(const std::vector<STOCK::KLinePoint>& klineData, int period1 /* = 6 */, int period2 /* = 14 */)
-{
-	std::vector<WRData> result;
-	result.reserve(klineData.size());
-	if (klineData.empty() || period1 <= 0 || period2 <= 0)
-		return result;
-
-	int maxPeriod = (std::max)(period1, period2);
-
-	for (int i = 0; i < static_cast<int>(klineData.size()); i++)
-	{
-		WRData wr;
-		wr.wr1 = 0;
-		wr.wr2 = 0;
-		wr.valid = false;
-
-		// 计算WR1（短期，如N=6）—— 使用K线的high/low
-		if (i >= period1 - 1)
-		{
-			STOCK::Price highest = 0;
-			STOCK::Price lowest = (std::numeric_limits<STOCK::Price>::max)();
-			bool hasValid = false;
-			for (int j = i - period1 + 1; j <= i; j++)
-			{
-				if (klineData[j].high > 0 && klineData[j].low > 0)
-				{
-					highest = (std::max)(highest, klineData[j].high);
-					lowest = (std::min)(lowest, klineData[j].low);
-					hasValid = true;
-				}
-			}
-			if (hasValid && highest > lowest && klineData[i].close > 0)
-			{
-				wr.wr1 = (static_cast<double>(highest) - klineData[i].close) / (highest - lowest) * 100.0;
-				if (wr.wr1 < 0) wr.wr1 = 0;
-				if (wr.wr1 > 100) wr.wr1 = 100;
-			}
-			else if (hasValid && highest == lowest && klineData[i].close > 0)
-			{
-				wr.wr1 = 0;
-			}
-		}
-
-		// 计算WR2（长期，如N=14）—— 使用K线的high/low
-		if (i >= period2 - 1)
-		{
-			STOCK::Price highest = 0;
-			STOCK::Price lowest = (std::numeric_limits<STOCK::Price>::max)();
-			bool hasValid = false;
-			for (int j = i - period2 + 1; j <= i; j++)
-			{
-				if (klineData[j].high > 0 && klineData[j].low > 0)
-				{
-					highest = (std::max)(highest, klineData[j].high);
-					lowest = (std::min)(lowest, klineData[j].low);
-					hasValid = true;
-				}
-			}
-			if (hasValid && highest > lowest && klineData[i].close > 0)
-			{
-				wr.wr2 = (static_cast<double>(highest) - klineData[i].close) / (highest - lowest) * 100.0;
-				if (wr.wr2 < 0) wr.wr2 = 0;
-				if (wr.wr2 > 100) wr.wr2 = 100;
-			}
-			else if (hasValid && highest == lowest && klineData[i].close > 0)
-			{
-				wr.wr2 = 0;
-			}
-		}
-
-		wr.valid = (i >= maxPeriod - 1);
-		result.push_back(wr);
-	}
-
-	return result;
-}
+// ========== W&R威廉指标绘制 ==========
+// 注：CalculateTimelineWR/CalculateKLineWR 已移至 CStockIndicator 类。
 
 void CFloatingWnd::DrawTimelineWRChart(CDC& memDC, int x, int y, int width, int height, const std::vector<STOCK::TimelinePoint>& timelinePoint, const std::vector<WRData>& wrData, int startIndex /* = 0 */, int xAxisPoints /* = 0 */)
 {
@@ -5032,19 +4277,19 @@ void CFloatingWnd::DrawTimelineWRSection(CDC& memDC, const TimelineDrawContext& 
 	std::vector<WRData> wrData;
 	if ((m_isMin5KLineMode || m_isMin30KLineMode) && ctx.klineData)
 	{
-		wrData = CalculateKLineWR(*ctx.klineData);
+		wrData = CStockIndicator::CalculateKLineWR(*ctx.klineData);
 	}
 	else if (m_isKLineMode && !m_isMin5KLineMode && !m_isMin30KLineMode)
 	{
 		// 日K线模式使用klineData
 		if (ctx.klineData)
-			wrData = CalculateKLineWR(*ctx.klineData);
+			wrData = CStockIndicator::CalculateKLineWR(*ctx.klineData);
 	}
 	else
 	{
 		// 分时模式使用完整分时数据计算，避免缩放/拖动时前N-1个点无效
 		const auto& fullData = ctx.fullTimeline ? *ctx.fullTimeline : timelinePoint;
-		wrData = CalculateTimelineWR(fullData);
+		wrData = CStockIndicator::CalculateTimelineWR(fullData);
 	}
 
 	DrawTimelineWRChart(memDC, 0, ctx.macdChartTop, ctx.chartWidth, ctx.macdChartHeight, timelinePoint, wrData, ctx.startIndex, ctx.xAxisPoints);
@@ -5091,144 +4336,8 @@ void CFloatingWnd::DrawTimelineWRSection(CDC& memDC, const TimelineDrawContext& 
 	}
 }
 
-// ========== RSI相对强弱指标 ==========
-
-std::vector<CFloatingWnd::RSIData> CFloatingWnd::CalculateTimelineRSI(const std::vector<STOCK::TimelinePoint>& timelinePoint, int period1 /* = 6 */, int period2 /* = 14 */)
-{
-	std::vector<RSIData> result;
-	result.reserve(timelinePoint.size());
-	if (timelinePoint.size() < 2 || period1 <= 1 || period2 <= 1)
-		return result;
-
-	int maxPeriod = (std::max)(period1, period2);
-
-	// 辅助lambda：计算指定周期的RSI序列
-	auto calcRSI = [&](int period) -> std::vector<double> {
-		std::vector<double> rsiValues(timelinePoint.size(), 0);
-		// 初始AU/AD：前period根K线的平均涨跌幅
-		double au = 0, ad = 0;
-		int count = 0;
-		for (int i = 1; i <= period && i < static_cast<int>(timelinePoint.size()); i++)
-		{
-			double delta = timelinePoint[i].price - timelinePoint[i - 1].price;
-			if (delta > 0)
-				au += delta;
-			else
-				ad += (-delta);
-			count++;
-		}
-		if (count > 0)
-		{
-			au /= count;
-			ad /= count;
-		}
-
-		// 第period根的RSI
-		if (period < static_cast<int>(timelinePoint.size()))
-		{
-			if (au == 0 && ad == 0)
-				rsiValues[period] = 50.0;
-			else
-				rsiValues[period] = 100.0 - 100.0 / (1.0 + au / (ad > 0 ? ad : 1e-10));
-		}
-
-		// 从第period+1根开始，使用SMA平滑
-		for (int i = period + 1; i < static_cast<int>(timelinePoint.size()); i++)
-		{
-			double delta = timelinePoint[i].price - timelinePoint[i - 1].price;
-			double up = (delta > 0) ? delta : 0.0;
-			double dn = (delta < 0) ? (-delta) : 0.0;
-			au = (au * (period - 1) + up) / period;
-			ad = (ad * (period - 1) + dn) / period;
-			if (au == 0 && ad == 0)
-				rsiValues[i] = 50.0;
-			else
-				rsiValues[i] = 100.0 - 100.0 / (1.0 + au / (ad > 0 ? ad : 1e-10));
-		}
-		return rsiValues;
-		};
-
-	auto rsi1Values = calcRSI(period1);
-	auto rsi2Values = calcRSI(period2);
-
-	for (int i = 0; i < static_cast<int>(timelinePoint.size()); i++)
-	{
-		RSIData rd;
-		rd.rsi1 = rsi1Values[i];
-		rd.rsi2 = rsi2Values[i];
-		rd.valid = (i >= maxPeriod);
-		result.push_back(rd);
-	}
-
-	return result;
-}
-
-std::vector<CFloatingWnd::RSIData> CFloatingWnd::CalculateKLineRSI(const std::vector<STOCK::KLinePoint>& klineData, int period1 /* = 6 */, int period2 /* = 14 */)
-{
-	std::vector<RSIData> result;
-	result.reserve(klineData.size());
-	if (klineData.size() < 2 || period1 <= 1 || period2 <= 1)
-		return result;
-
-	int maxPeriod = (std::max)(period1, period2);
-
-	// 辅助lambda：计算指定周期的RSI序列
-	auto calcRSI = [&](int period) -> std::vector<double> {
-		std::vector<double> rsiValues(klineData.size(), 0);
-		double au = 0, ad = 0;
-		int count = 0;
-		for (int i = 1; i <= period && i < static_cast<int>(klineData.size()); i++)
-		{
-			double delta = klineData[i].close - klineData[i - 1].close;
-			if (delta > 0)
-				au += delta;
-			else
-				ad += (-delta);
-			count++;
-		}
-		if (count > 0)
-		{
-			au /= count;
-			ad /= count;
-		}
-
-		if (period < static_cast<int>(klineData.size()))
-		{
-			if (au == 0 && ad == 0)
-				rsiValues[period] = 50.0;
-			else
-				rsiValues[period] = 100.0 - 100.0 / (1.0 + au / (ad > 0 ? ad : 1e-10));
-		}
-
-		for (int i = period + 1; i < static_cast<int>(klineData.size()); i++)
-		{
-			double delta = klineData[i].close - klineData[i - 1].close;
-			double up = (delta > 0) ? delta : 0.0;
-			double dn = (delta < 0) ? (-delta) : 0.0;
-			au = (au * (period - 1) + up) / period;
-			ad = (ad * (period - 1) + dn) / period;
-			if (au == 0 && ad == 0)
-				rsiValues[i] = 50.0;
-			else
-				rsiValues[i] = 100.0 - 100.0 / (1.0 + au / (ad > 0 ? ad : 1e-10));
-		}
-		return rsiValues;
-		};
-
-	auto rsi1Values = calcRSI(period1);
-	auto rsi2Values = calcRSI(period2);
-
-	for (int i = 0; i < static_cast<int>(klineData.size()); i++)
-	{
-		RSIData rd;
-		rd.rsi1 = rsi1Values[i];
-		rd.rsi2 = rsi2Values[i];
-		rd.valid = (i >= maxPeriod);
-		result.push_back(rd);
-	}
-
-	return result;
-}
+// ========== RSI相对强弱指标绘制 ==========
+// 注：CalculateTimelineRSI/CalculateKLineRSI 已移至 CStockIndicator 类。
 
 void CFloatingWnd::DrawTimelineRSIChart(CDC& memDC, int x, int y, int width, int height, const std::vector<STOCK::TimelinePoint>& timelinePoint, const std::vector<RSIData>& rsiData, int startIndex /* = 0 */, int xAxisPoints /* = 0 */)
 {
@@ -5302,17 +4411,17 @@ void CFloatingWnd::DrawTimelineRSISection(CDC& memDC, const TimelineDrawContext&
 	std::vector<RSIData> rsiData;
 	if ((m_isMin5KLineMode || m_isMin30KLineMode) && ctx.klineData)
 	{
-		rsiData = CalculateKLineRSI(*ctx.klineData);
+		rsiData = CStockIndicator::CalculateKLineRSI(*ctx.klineData);
 	}
 	else if (m_isKLineMode && !m_isMin5KLineMode && !m_isMin30KLineMode)
 	{
 		if (ctx.klineData)
-			rsiData = CalculateKLineRSI(*ctx.klineData);
+			rsiData = CStockIndicator::CalculateKLineRSI(*ctx.klineData);
 	}
 	else
 	{
 		const auto& fullData = ctx.fullTimeline ? *ctx.fullTimeline : timelinePoint;
-		rsiData = CalculateTimelineRSI(fullData);
+		rsiData = CStockIndicator::CalculateTimelineRSI(fullData);
 	}
 
 	DrawTimelineRSIChart(memDC, 0, ctx.macdChartTop, ctx.chartWidth, ctx.macdChartHeight, timelinePoint, rsiData, ctx.startIndex, ctx.xAxisPoints);
@@ -5477,30 +4586,7 @@ CFloatingWnd::KLineDrawData CFloatingWnd::PrepareKLineDrawData(int x, int y, int
 	return d;
 }
 
-void CFloatingWnd::CalculatePeriodHighsLows(const KLineDrawData& drawData, PeriodPoint periodHighs[3], PeriodPoint periodLows[3], bool useClose)
-{
-	const auto& klineData = *drawData.klineData;
-	const int DAYS_PER_YEAR = 250;
-
-	for (int p = 1; p <= 3; p++)
-	{
-		int rangeEnd = klineData.size() - (p - 1) * DAYS_PER_YEAR;
-		int rangeStart = max(drawData.startIndex, static_cast<int>(klineData.size()) - p * DAYS_PER_YEAR);
-		if (rangeStart >= rangeEnd) continue;
-
-		STOCK::Price hh = 0, ll = (std::numeric_limits<STOCK::Price>::max)();
-		int hIdx = -1, lIdx = -1;
-		for (int i = rangeStart; i < rangeEnd; i++)
-		{
-			STOCK::Price price = useClose ? klineData[i].close : klineData[i].high;
-			STOCK::Price lowPrice = useClose ? klineData[i].close : klineData[i].low;
-			if (price > 0 && price > hh) { hh = price; hIdx = i; }
-			if (lowPrice > 0 && lowPrice < ll) { ll = lowPrice; lIdx = i; }
-		}
-		periodHighs[p - 1] = { hIdx, hh, hIdx >= 0 ? klineData[hIdx].day : "" };
-		periodLows[p - 1] = { lIdx, ll, lIdx >= 0 ? klineData[lIdx].day : "" };
-	}
-}
+// 注：CalculatePeriodHighsLows 已移至 CStockIndicator 类。
 
 std::vector<CFloatingWnd::LabelInfo> CFloatingWnd::DrawKLineMonthLines(CDC& memDC, const KLineDrawData& drawData)
 {
@@ -5996,7 +5082,7 @@ void CFloatingWnd::DrawKLineChart(CDC& memDC, int x, int y, int w, int h, const 
 	// 计算周期高低点（仅用于标记显示）
 	PeriodPoint periodHighs[3] = {};
 	PeriodPoint periodLows[3] = {};
-	CalculatePeriodHighsLows(drawData, periodHighs, periodLows);
+	CStockIndicator::CalculatePeriodHighsLows(*drawData.klineData, drawData.startIndex, periodHighs, periodLows);
 
 	// Y轴固定6等分7根横线：以可见最高/最低价的中点为Y轴中线，向上下对称扩展（与5分钟K线一致）
 	if (drawData.maxPrice > drawData.minPrice)
@@ -6004,7 +5090,7 @@ void CFloatingWnd::DrawKLineChart(CDC& memDC, int x, int y, int w, int h, const 
 		const double DIV_COUNT = 6.0;
 		const double MIN_STEP = 0.001;
 		double niceMin, niceMax, niceStep;
-		CalcNiceAxisRangeSymmetric(drawData.minPrice, drawData.maxPrice, DIV_COUNT, MIN_STEP, niceMin, niceMax, niceStep);
+		CStockIndicator::CalcNiceAxisRangeSymmetric(drawData.minPrice, drawData.maxPrice, DIV_COUNT, MIN_STEP, niceMin, niceMax, niceStep);
 		drawData.minPrice = niceMin;
 		drawData.maxPrice = niceMax;
 		drawData.unitY = drawData.h / (drawData.maxPrice - drawData.minPrice);
@@ -6374,7 +5460,7 @@ void CFloatingWnd::DrawKLineTrendChart(CDC& memDC, int x, int y, int w, int h, c
 
 	PeriodPoint periodHighs[3] = {};
 	PeriodPoint periodLows[3] = {};
-	CalculatePeriodHighsLows(drawData, periodHighs, periodLows, true);
+	CStockIndicator::CalculatePeriodHighsLows(*drawData.klineData, drawData.startIndex, periodHighs, periodLows, true);
 
 	// Y轴固定6等分7根横线：以可见最高/最低价的中点为Y轴中线，向上下对称扩展（与5分钟K线一致）
 	if (drawData.maxPrice > drawData.minPrice)
@@ -6382,7 +5468,7 @@ void CFloatingWnd::DrawKLineTrendChart(CDC& memDC, int x, int y, int w, int h, c
 		const double DIV_COUNT = 6.0;
 		const double MIN_STEP = 0.001;
 		double niceMin, niceMax, niceStep;
-		CalcNiceAxisRangeSymmetric(drawData.minPrice, drawData.maxPrice, DIV_COUNT, MIN_STEP, niceMin, niceMax, niceStep);
+		CStockIndicator::CalcNiceAxisRangeSymmetric(drawData.minPrice, drawData.maxPrice, DIV_COUNT, MIN_STEP, niceMin, niceMax, niceStep);
 		drawData.minPrice = niceMin;
 		drawData.maxPrice = niceMax;
 		drawData.unitY = drawData.h / (drawData.maxPrice - drawData.minPrice);
@@ -9048,173 +8134,15 @@ void CFloatingWnd::OnLButtonUp(UINT nFlags, CPoint point)
 	CWnd::OnLButtonUp(nFlags, point);
 }
 
-// ========== KDJ 指标计算与绘制 ==========
-
-std::vector<CFloatingWnd::KDJData> CFloatingWnd::CalculateKDJ(const std::vector<STOCK::KLinePoint>& klineData, int period /* = 9 */)
-{
-	std::vector<KDJData> result;
-	result.reserve(klineData.size());
-	if (klineData.empty() || period <= 0)
-		return result;
-
-	// 初始 K = D = 50
-	double prevK = 50.0;
-	double prevD = 50.0;
-
-	for (int i = 0; i < static_cast<int>(klineData.size()); i++)
-	{
-		KDJData kd;
-		kd.k = 0;
-		kd.d = 0;
-		kd.j = 0;
-		kd.valid = false;
-
-		// 需要至少 period 个数据点才能计算 RSV
-		if (i >= period - 1)
-		{
-			STOCK::Price highest = 0;
-			STOCK::Price lowest = (std::numeric_limits<STOCK::Price>::max)();
-			bool hasValid = false;
-			for (int j = i - period + 1; j <= i; j++)
-			{
-				if (klineData[j].high > 0)
-				{
-					highest = (std::max)(highest, klineData[j].high);
-					hasValid = true;
-				}
-				if (klineData[j].low > 0)
-				{
-					lowest = (std::min)(lowest, klineData[j].low);
-					hasValid = true;
-				}
-			}
-
-			if (hasValid && highest > lowest && klineData[i].close > 0)
-			{
-				double rsv = (static_cast<double>(klineData[i].close) - lowest) / (highest - lowest) * 100.0;
-				// 限制 RSV 在 [0, 100] 范围
-				if (rsv < 0) rsv = 0;
-				if (rsv > 100) rsv = 100;
-
-				// K(t) = 2/3 * K(t-1) + 1/3 * RSV(t)
-				double curK = 2.0 / 3.0 * prevK + 1.0 / 3.0 * rsv;
-				// D(t) = 2/3 * D(t-1) + 1/3 * K(t)
-				double curD = 2.0 / 3.0 * prevD + 1.0 / 3.0 * curK;
-				// J = 3K - 2D
-				double curJ = 3.0 * curK - 2.0 * curD;
-
-				kd.k = curK;
-				kd.d = curD;
-				kd.j = curJ;
-				kd.valid = true;
-
-				prevK = curK;
-				prevD = curD;
-			}
-			else
-			{
-				// 数据无效，保持上一状态
-				kd.k = prevK;
-				kd.d = prevD;
-				kd.j = 3.0 * prevK - 2.0 * prevD;
-				kd.valid = true;
-			}
-		}
-		else
-		{
-			// 前 period-1 个数据点无法计算 RSV，使用初始值 50
-			kd.k = prevK;
-			kd.d = prevD;
-			kd.j = 3.0 * prevK - 2.0 * prevD;
-			kd.valid = true;
-		}
-
-		result.push_back(kd);
-	}
-
-	return result;
-}
-
-std::vector<CFloatingWnd::KDJData> CFloatingWnd::CalculateTimelineKDJ(const std::vector<STOCK::TimelinePoint>& timelinePoint, int period /* = 9 */)
-{
-	std::vector<KDJData> result;
-	result.reserve(timelinePoint.size());
-	if (timelinePoint.empty() || period <= 0)
-		return result;
-
-	double prevK = 50.0;
-	double prevD = 50.0;
-
-	for (int i = 0; i < static_cast<int>(timelinePoint.size()); i++)
-	{
-		KDJData kd;
-		kd.k = 0;
-		kd.d = 0;
-		kd.j = 0;
-		kd.valid = false;
-
-		if (i >= period - 1)
-		{
-			STOCK::Price highest = 0;
-			STOCK::Price lowest = (std::numeric_limits<STOCK::Price>::max)();
-			bool hasValid = false;
-			for (int j = i - period + 1; j <= i; j++)
-			{
-				STOCK::Price price = timelinePoint[j].price;
-				if (price > 0)
-				{
-					highest = (std::max)(highest, price);
-					lowest = (std::min)(lowest, price);
-					hasValid = true;
-				}
-			}
-
-			if (hasValid && highest > lowest && timelinePoint[i].price > 0)
-			{
-				double rsv = (static_cast<double>(timelinePoint[i].price) - lowest) / (highest - lowest) * 100.0;
-				if (rsv < 0) rsv = 0;
-				if (rsv > 100) rsv = 100;
-
-				double curK = 2.0 / 3.0 * prevK + 1.0 / 3.0 * rsv;
-				double curD = 2.0 / 3.0 * prevD + 1.0 / 3.0 * curK;
-				double curJ = 3.0 * curK - 2.0 * curD;
-
-				kd.k = curK;
-				kd.d = curD;
-				kd.j = curJ;
-				kd.valid = true;
-
-				prevK = curK;
-				prevD = curD;
-			}
-			else
-			{
-				kd.k = prevK;
-				kd.d = prevD;
-				kd.j = 3.0 * prevK - 2.0 * prevD;
-				kd.valid = true;
-			}
-		}
-		else
-		{
-			kd.k = prevK;
-			kd.d = prevD;
-			kd.j = 3.0 * prevK - 2.0 * prevD;
-			kd.valid = true;
-		}
-
-		result.push_back(kd);
-	}
-
-	return result;
-}
+// ========== KDJ 指标绘制 ==========
+// 注：CalculateKDJ/CalculateTimelineKDJ 已移至 CStockIndicator 类。
 
 void CFloatingWnd::DrawKDJChart(CDC& memDC, int x, int y, int width, int height, const std::vector<STOCK::KLinePoint>& klineData)
 {
 	if (klineData.empty())
 		return;
 
-	auto kdjData = CalculateKDJ(klineData);
+	auto kdjData = CStockIndicator::CalculateKDJ(klineData);
 	if (kdjData.empty())
 		return;
 
@@ -10594,7 +9522,7 @@ void CFloatingWnd::OnMouseMove(UINT nFlags, CPoint point)
 			int startIndex = max(0, min(m_timelineScrollOffset, maxOffset));
 
 			// 构建可见子向量并计算MA值（与OnPaint一致）
-			CalcAllRollingAvgPrices(timelinePoint);
+			CStockIndicator::CalcAllRollingAvgPrices(timelinePoint);
 			auto subStart = timelinePoint.begin() + startIndex;
 			auto subEnd = timelinePoint.begin() + startIndex + visibleCount;
 			std::vector<STOCK::TimelinePoint> subTimeline(subStart, subEnd);
@@ -10603,7 +9531,9 @@ void CFloatingWnd::OnMouseMove(UINT nFlags, CPoint point)
 			int adjX = point.x - chartLeft;
 			int effectiveWidth = chartWidth - chartLeft;
 			// 按索引比例计算鼠标对应的可见数据索引
-			int relIndex = static_cast<int>(adjX * static_cast<float>(visibleCount) / effectiveWidth);
+			// 分时模式X轴基于m_timelineVisibleCount固定格数，K线模式基于实际数据点数
+			int xSlotCount = m_isKLineMode ? visibleCount : m_timelineVisibleCount;
+			int relIndex = static_cast<int>(adjX * static_cast<float>(xSlotCount) / effectiveWidth);
 			relIndex = max(0, min(relIndex, visibleCount - 1));
 
 			if (relIndex >= 0 && relIndex < static_cast<int>(subTimeline.size()))
@@ -10641,7 +9571,7 @@ void CFloatingWnd::OnMouseMove(UINT nFlags, CPoint point)
 				// 设置MACD/KDJ/W&R/RSI标题栏悬停提示
 				if (m_timelineIndicator == TimelineIndicator::KDJ)
 				{
-					auto kdjData = CalculateTimelineKDJ(subTimeline);
+					auto kdjData = CStockIndicator::CalculateTimelineKDJ(subTimeline);
 					if (relIndex < static_cast<int>(kdjData.size()) && kdjData[relIndex].valid)
 					{
 						m_timelineKdjTitleTip.Format(_T("K:%.1f D:%.1f J:%.1f"), kdjData[relIndex].k, kdjData[relIndex].d, kdjData[relIndex].j);
@@ -10653,7 +9583,7 @@ void CFloatingWnd::OnMouseMove(UINT nFlags, CPoint point)
 				else if (m_timelineIndicator == TimelineIndicator::WR)
 				{
 					// WR悬停提示
-					auto wrData = CalculateTimelineWR(subTimeline);
+					auto wrData = CStockIndicator::CalculateTimelineWR(subTimeline);
 					if (relIndex < static_cast<int>(wrData.size()) && wrData[relIndex].valid)
 					{
 						m_timelineWrTitleTip.Format(_T("WR6:%.1f WR14:%.1f"), wrData[relIndex].wr1, wrData[relIndex].wr2);
@@ -10664,7 +9594,7 @@ void CFloatingWnd::OnMouseMove(UINT nFlags, CPoint point)
 				}
 				else if (m_timelineIndicator == TimelineIndicator::RSI)
 				{
-					auto rsiData = CalculateTimelineRSI(subTimeline);
+					auto rsiData = CStockIndicator::CalculateTimelineRSI(subTimeline);
 					if (relIndex < static_cast<int>(rsiData.size()) && rsiData[relIndex].valid)
 					{
 						m_timelineRsiTitleTip.Format(_T("RSI6:%.1f RSI14:%.1f"), rsiData[relIndex].rsi1, rsiData[relIndex].rsi2);
@@ -10675,7 +9605,7 @@ void CFloatingWnd::OnMouseMove(UINT nFlags, CPoint point)
 				}
 				else
 				{
-					auto macdData = CalculateTimelineMACD(subTimeline);
+					auto macdData = CStockIndicator::CalculateTimelineMACD(subTimeline);
 					if (relIndex < static_cast<int>(macdData.size()) && macdData[relIndex].valid)
 					{
 						m_timelineMacdTitleTip.Format(_T("DIF:%.3f DEA:%.3f MACD:%.3f"), macdData[relIndex].dif, macdData[relIndex].dea, macdData[relIndex].bar);
