+++ Plugins/Stock/Stock.vcxproj.filters
@@ -69,6 +69,9 @@
     <ClInclude Include="SignalAnalyzer.h">
       <Filter>SignalAnalyzer</Filter>
     </ClInclude>
+    <ClInclude Include="StockIndicator.h">
+      <Filter>SignalAnalyzer</Filter>
+    </ClInclude>
     <ClInclude Include="StockFetchThread.h">
       <Filter>Stock</Filter>
     </ClInclude>
@@ -120,6 +123,9 @@
       <Filter>FloatingWnd</Filter>
     </ClCompile>
     <ClCompile Include="SignalAnalyzer.cpp">
+      <Filter>SignalAnalyzer</Filter>
+    </ClCompile>
+    <ClCompile Include="StockIndicator.cpp">
       <Filter>SignalAnalyzer</Filter>
     </ClCompile>
     <ClCompile Include="StockFetchThread.cpp">
